added gitsend alias
This is used to ensure no test-failing commits are pushed.
(Using `make test` to be language-agnostic. This happens to also reveal
repos which are missing a makefile with a test target.)

Using shell alias because Git doesn't have a pre-commit hook and Git
aliases' CWD is the repository root, which isn't always suitable.